Import of AUTRIJUS/PAR-0.79_99 from CPAN.
gitpan-cpan-distribution: PAR
gitpan-cpan-version:      0.79_99
gitpan-cpan-path:         AUTRIJUS/PAR-0.79_99.tar.gz
gitpan-cpan-author:       AUTRIJUS
gitpan-cpan-maturity:     developer

+++ myldr/mktmpdir.c
@@ -1,5 +1,5 @@
 /* $File: //member/autrijus/PAR/myldr/mktmpdir.c $ $Author: autrijus $
-   $Revision: #31 $ $Change: 10199 $ $DateTime: 2004/02/24 18:14:30 $
+   $Revision: #34 $ $Change: 10300 $ $DateTime: 2004/03/03 04:07:09 $
    vim: expandtab shiftwidth=4
 */
 
@@ -110,7 +110,7 @@ char *par_mktmpdir ( char **argv ) {
     else {
         /* "$TEMP/par-$USER/temp-$PID" */
 
-        putenv("PAR_CLEAN=1");
+        par_setenv("PAR_CLEAN", "1");
         sprintf(
             stmpdir,
             "%s%stemp-%u%s",
@@ -120,30 +120,27 @@ char *par_mktmpdir ( char **argv ) {
 
     /* set dynamic loading path */
     par_temp_env = (char *)malloc(strlen(PAR_TEMP) + strlen(stmpdir) + 2);
-    sprintf(par_temp_env, "%s=%s", PAR_TEMP, stmpdir);
-    putenv(par_temp_env);
+    par_setenv(PAR_TEMP, stmpdir);
 
     for ( i = 0 ; strlen(key = ld_path_keys[i]) > 0 ; i++ ) {
         if ( (val = (char *)getenv(key)) == NULL ) continue;
 
         if ( strlen(val) == 0 ) {
-            ld_path_env = (char *)malloc(strlen(key) + strlen(stmpdir) + 2);
-            sprintf(ld_path_env, "%s=%s", key, stmpdir);
+            par_setenv(key, stmpdir);
         }
         else {
             ld_path_env = (char *)malloc(
-                strlen(key) +
                 strlen(stmpdir) +
                 strlen(path_sep) +
                 strlen(val) + 2
             );
             sprintf(
                 ld_path_env,
-                "%s=%s%s%s",
-                key, stmpdir, path_sep, val
+                "%s%s%s",
+                stmpdir, path_sep, val
             );
+            par_setenv(key, ld_path_env);
         }
-        putenv(ld_path_env);
     }
 
     return(stmpdir);
@@ -151,9 +148,10 @@ char *par_mktmpdir ( char **argv ) {
 
 
 #ifdef WIN32
-void par_rmtmpdir ( char *stmpdir ) {
+void par_rmtmpdir ( char *stmpdir, int recurse ) {
     struct _finddata_t cur_file;
     char *subsubdir = malloc(strlen(stmpdir) + 258);
+    char *slashdot;
     long hFile;
 	int tries = 0;
     HMODULE dll;
@@ -171,8 +169,19 @@ void par_rmtmpdir ( char *stmpdir ) {
         sprintf(subsubdir, "%s", cur_file.name);
     }
 
-    /* if (!(cur_file.attrib & _A_SUBDIR)) fprintf(stderr, "unlinking %s\n", subsubdir); */
-    if (!(cur_file.attrib & _A_SUBDIR)) _unlink(subsubdir);
+    if (!(slashdot = strstr(subsubdir, "\\.")) || (strcmp(slashdot,"\\.") && strcmp(slashdot,"\\.."))) {
+        if ((cur_file.attrib & _A_SUBDIR) && recurse) {
+            par_rmtmpdir( subsubdir, 1 );
+        }
+        /* if (!(cur_file.attrib & _A_SUBDIR)) fprintf(stderr, "unlinking %s\n", subsubdir); */
+        else {
+            dll = GetModuleHandle(cur_file.name);
+            tries = 0;
+            while ( _unlink(subsubdir) && ( tries++ < 10 ) ) {
+                if ( dll ) FreeLibrary(dll);
+            };
+        }
+    }
     while ( _findnext( hFile, &cur_file ) == 0 ) {
         if (!strstr(cur_file.name, "\\")) {
             sprintf(subsubdir, "%s\\%s", stmpdir, cur_file.name);
@@ -181,13 +190,18 @@ void par_rmtmpdir ( char *stmpdir ) {
             sprintf(subsubdir, "%s", cur_file.name);
         }
 
-        /* if (!(cur_file.attrib & _A_SUBDIR)) fprintf(stderr, "unlinking %s\n", subsubdir); */
-        if (!(cur_file.attrib & _A_SUBDIR)) {
-            dll = GetModuleHandle(cur_file.name);
-            tries = 0;
-            while ( _unlink(subsubdir) && ( tries++ < 10 ) ) {
-                if ( dll ) FreeLibrary(dll);
-            };
+        if (!(slashdot = strstr(subsubdir, "\\.")) || (strcmp(slashdot,"\\.") && strcmp(slashdot,"\\.."))) {
+            if ((cur_file.attrib & _A_SUBDIR) && recurse) {
+                par_rmtmpdir( subsubdir, 1 );
+            }
+            /* if (!(cur_file.attrib & _A_SUBDIR)) fprintf(stderr, "unlinking %s\n", subsubdir); */
+            else {
+                dll = GetModuleHandle(cur_file.name);
+                tries = 0;
+                while ( _unlink(subsubdir) && ( tries++ < 10 ) ) {
+                    if ( dll ) FreeLibrary(dll);
+                };
+            }
         }
     }
 
@@ -196,10 +210,11 @@ void par_rmtmpdir ( char *stmpdir ) {
 }
 
 #else
-void par_rmtmpdir ( char *stmpdir ) {
+void par_rmtmpdir ( char *stmpdir, int recurse ) {
     DIR *partmp_dirp;
     Direntry_t *dp;
     char *subsubdir;
+    struct stat stbuf;
 
     /* remove temporary PAR directory */
     partmp_dirp = opendir(stmpdir);
@@ -207,13 +222,17 @@ void par_rmtmpdir ( char *stmpdir ) {
     if ( partmp_dirp == NULL ) return;
 
     /* fprintf(stderr, "%s: removing private temporary subdirectory %s.\n", argv[0], stmpdir); */
-    /* here we simply assume that PAR will NOT create any subdirectories ... */
     while ( ( dp = readdir(partmp_dirp) ) != NULL ) {
         if ( strcmp (dp->d_name, ".") != 0 && strcmp (dp->d_name, "..") != 0 )
         {
             subsubdir = malloc(strlen(stmpdir) + strlen(dp->d_name) + 2);
             sprintf(subsubdir, "%s/%s", stmpdir, dp->d_name);
-            unlink(subsubdir);
+            if (stat(subsubdir, &stbuf) != -1 && S_ISDIR(stbuf.st_mode) && recurse) {
+                par_rmtmpdir(subsubdir, 1);
+            }
+            else {
+                unlink(subsubdir);
+            }
             free(subsubdir);
             subsubdir = NULL;
         }
@@ -229,9 +248,8 @@ void par_cleanup (char *stmpdir) {
     char *basename = par_basename(dirname);
     if ( par_env_clean() && stmpdir != NULL && strlen(stmpdir)) {
         if ( strstr(basename, "par-") == basename ) {
-            par_rmtmpdir(stmpdir);
-            par_rmtmpdir(dirname);
+            par_rmtmpdir(stmpdir, 1);
+            par_rmtmpdir(dirname, 0);
         }
     }
 }
-
